+++ renderer/core-json-renderer/src/main/groovy/com/ainrif/apiator/renderer/core/json/plugin/modeltype/DefaultModelTypeCompositePlugin.groovy
@@ -27,7 +27,8 @@ class DefaultModelTypeCompositePlugin implements CompositePlugin {
                 new CollectionsModelTypePlugin(),
                 new BinaryModelTypePlugin(),
                 new OldDateModelTypePlugin(),
-                new CoreJavaModelTypePlugin()
+                new CoreJavaModelTypePlugin(),
+                new Jsr310ModelTypePlugin(),
         ]
     }
 }