+++ renderer/core-json-renderer/src/main/groovy/com/ainrif/apiator/renderer/core/json/plugin/modeltype/Jsr310ModelTypePlugin.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ainrif.apiator.renderer.plugin.j8
+package com.ainrif.apiator.renderer.core.json.plugin.modeltype
 
 import com.ainrif.apiator.renderer.plugin.spi.modeltype.ModelType
 import com.ainrif.apiator.renderer.plugin.spi.modeltype.ModelTypePlugin