+++ renderer/core-json-renderer/src/test/groovy/com/ainrif/apiator/renderer/core/json/plugin/modeltype/Jsr310ModelTypePluginSpec.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ainrif.apiator.renderer.plugin.j8
+package com.ainrif.apiator.renderer.core.json.plugin.modeltype
 
 import com.ainrif.apiator.renderer.plugin.spi.modeltype.ModelType
 import spock.lang.Specification