+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.9.x
 - new: micronaut provider
 - new: path transformation plugin
 - improvement: reduce amount of file scans for javadoc 
+- improvement: jsr310 plugin becomes a core plugin. jdk8+ required
 
 v0.8.x - Renderer Plugins